changed noConnection from firestore failure to state
removed noConnection firestore failure and switched to state.
displays text when state is noConnection
