+++ lib/consumer/application/retailer_list_state.dart
@@ -7,6 +7,7 @@ part 'retailer_list_state.freezed.dart';
 class RetailerListState with _$RetailerListState {
   const RetailerListState._();
   const factory RetailerListState.initial() = Initial;
+  const factory RetailerListState.noConnection() = NoConnection;
   const factory RetailerListState.loading() = Loading;
   const factory RetailerListState.loaded(List<Retailer> retailers) = Loaded;
   const factory RetailerListState.failure(String message) = Failure;
