+++ lib/consumer/infrastructure/firestore_failures.dart
@@ -7,6 +7,5 @@ class FirestoreFailures with _$FirestoreFailures {
   const FirestoreFailures._();
   const factory FirestoreFailures.cancelledOperation() = CancelledOperation;
   const factory FirestoreFailures.objectNotFound() = ObjectNotFound;
-  const factory FirestoreFailures.noConnection() = NoConnection;
   const factory FirestoreFailures.unknown() = Unknown;
 }
